+++ src/screens/SupportChatScreen.tsx
@@ -1,211 +1,77 @@
-import React, { useCallback, useEffect } from 'react';
+import React from 'react';
 import {
   StyleSheet,
   Text,
   View,
-  ActivityIndicator,
-  FlatList,
-  Image,
-  KeyboardAvoidingView,
+  Linking,
   Platform,
   TouchableOpacity,
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { useFocusEffect } from '@react-navigation/native';
-import firestore, {
-  FirebaseFirestoreTypes,
-} from '@react-native-firebase/firestore';
-import { AntDesign } from '@expo/vector-icons';
+import { Entypo, MaterialCommunityIcons } from '@expo/vector-icons';
 import { scale } from 'react-native-size-matters';
 import { dozy_theme } from '../config/Themes';
-import fetchChats from '../utilities/fetchChats';
-import sendChatMessage from '../utilities/sendChatMessage';
-import { ChatMessage } from '../components/ChatMessage';
-import { ChatTextInput } from '../components/ChatTextInput';
-import { Chat, Navigation } from '../types/custom';
-import Analytics from '../utilities/analytics.service';
-import AnalyticsEvents from '../constants/AnalyticsEvents';
 import FocusAwareStatusBar from '../components/FocusAwareStatusBar';
-import Auth from '../utilities/auth.service';
-import Notification from '../utilities/notification.service';
-import { useChatsStore } from '../utilities/chatsStore';
-import { useUserDataStore } from '../utilities/userDataStore';
+import Clipboard from '../../assets/images/Clipboard.svg';
+import { CardContainer } from '../components/CardContainer';
 
-export const SupportChatScreen: React.FC<{ navigation: Navigation }> = ({
-  navigation,
-}) => {
-  // Get global state
-  const { state } = Auth.useAuth();
-  const chats = useChatsStore((chatsState) => chatsState.chats);
-  const { coach, userData } = useUserDataStore((userDataState) => ({
-    coach: userDataState.coach,
-    userData: userDataState.userData,
-  }));
-  const setChatsInState = useChatsStore((chatsState) => chatsState.setChats);
-  const coachNameString = `${coach.firstName} ${coach.lastName}`;
-  // Set Firebase DB references if userId is defined
-  let colRef: FirebaseFirestoreTypes.CollectionReference;
-  if (state.userId) {
-    colRef = firestore()
-      .collection('users')
-      .doc(state.userId)
-      .collection('supportMessages');
-  }
-
-  useFocusEffect(
-    useCallback(() => {
-      // If LiveChat has a msg marked as unread, mark it as read in Firebase
-      if (userData?.livechatUnreadMsg) {
-        firestore().collection('users').doc(state.userId).update({
-          livechatUnreadMsg: false,
-        });
-        // Update the app icon's badge
-        Notification.calculateBadgeNumber(
-          state.userId!,
-          true,
-          true,
-          false,
-        ).then((badgeNumber) => {
-          Notification.setBadgeNumber(badgeNumber);
-        });
-      }
-    }, []),
-  );
-
-  // Function to fetch chats from Firebase and put them in global state
-  async function setChats() {
-    async function fetchData() {
-      if (!state.userId) throw new Error();
-      fetchChats(firestore(), state.userId)
-        .then((newChats: Array<Chat>) => {
-          // Check that theres >1 entry. If no, set state accordingly
-          if (newChats.length === 0) {
-            setChatsInState([]);
-          } else {
-            setChatsInState(newChats);
-          }
-
-          return;
-        })
-        .catch(function (error) {
-          console.log('Error getting chats:', error);
-        });
-    }
-
-    // Setup a listener to fetch new chats from Firebase
-    colRef.onSnapshot(function () {
-      fetchData();
-    });
-  }
-
-  // Set chats once upon loading
-  useEffect(() => {
-    setChats();
-    // Maybe remove CHAT_MESSAGE push notification
-    Notification.removeNotificationsFromTrayByType('CHAT_MESSAGE');
-  }, []);
-
+export const SupportChatScreen: React.FC = ({}) => {
   const theme = dozy_theme;
 
-  // If state is available, show screen. Otherwise, show loading indicator.
-  if (chats && userData?.currentTreatments) {
-    return (
-      <SafeAreaView style={styles.SafeAreaView} edges={['top']}>
-        <FocusAwareStatusBar backgroundColor={dozy_theme.colors.medium} />
-        <View style={styles.Root}>
-          <View style={styles.View_HeaderContainer}>
-            {!!coach.image && (
-              <Image source={coach.image} style={styles.Img_Profile} />
-            )}
-            <View style={styles.View_ChatNameContainer}>
-              <Text
-                style={{
-                  ...theme.typography.headline5,
-                  ...styles.Text_CoachName,
-                }}
-              >
-                {coachNameString}
-              </Text>
-              <Text
-                style={{ ...theme.typography.body2, ...styles.Text_CoachTitle }}
-              >
-                {coach.title}
-              </Text>
-            </View>
+  return (
+    <SafeAreaView style={styles.SafeAreaView} edges={['top']}>
+      <FocusAwareStatusBar backgroundColor={dozy_theme.colors.medium} />
+      <View style={styles.Root}>
+        <View style={styles.View_ContentContainer}>
+          <Clipboard
+            style={styles.Icon_Clipboard}
+            width={scale(80)}
+            height={scale(80)}
+          />
+          <CardContainer>
             <TouchableOpacity
-              onPress={() => navigation.navigate('TreatmentReview')}
-              style={styles.faqButton}
-            >
-              <AntDesign
-                name="questioncircle"
-                size={scale(25)}
-                color={theme.colors.secondary}
-              />
-              <Text style={{ ...theme.typography.body2, ...styles.Text_FAQ }}>
-                FAQ
-              </Text>
-            </TouchableOpacity>
-          </View>
-          <KeyboardAvoidingView
-            behavior={Platform.select({ ios: 'padding' })}
-            keyboardVerticalOffset={Platform.select({ ios: scale(60) })}
-            style={styles.KbAvoidingView}
-          >
-            <FlatList
-              contentContainerStyle={styles.View_ContentContainer}
-              renderItem={({ item }) => (
-                <ChatMessage
-                  message={item.message}
-                  time={(
-                    item.time as FirebaseFirestoreTypes.Timestamp
-                  ).toDate()}
-                  sentByUser={item.sentByUser}
-                  coach={coachNameString}
-                  pending={item.pending}
-                />
-              )}
-              keyExtractor={(item, index) => `${item.message}${index}`}
-              inverted={true}
-              data={chats}
-            />
-            <ChatTextInput
-              onSend={(typedMsg: string) => {
-                if (state.userId && typedMsg?.trim().length) {
-                  const newMessage: Chat = {
-                    sender:
-                      state.profileData.name || userData.userInfo?.displayName,
-                    message: typedMsg,
-                    time: firestore.Timestamp.fromDate(new Date()),
-                    sentByUser: true,
-                    pending: true,
-                  };
-                  setChatsInState([newMessage, ...chats]);
-                  sendChatMessage(firestore(), state.userId, newMessage);
-                  Analytics.logEvent(AnalyticsEvents.sendMessage);
-                }
+              onPress={() => {
+                Linking.openURL('https://discord.gg/KqDr4VmJKB');
               }}
-            />
-          </KeyboardAvoidingView>
+            >
+              <View style={styles.View_CardHeaderContainer}>
+                <View style={styles.headerContent}>
+                  <Text
+                    style={{
+                      ...theme.typography.cardTitle,
+                      ...styles.Text_CardTitle,
+                    }}
+                  >
+                    Have questions? Need help?
+                  </Text>
+                  <Text
+                    style={{
+                      ...theme.typography.body2,
+                      ...styles.Text_CardSubtitle,
+                    }}
+                  >
+                    Join our community discord server
+                  </Text>
+                </View>
+                <Entypo
+                  name={'chevron-right'}
+                  size={scale(28)}
+                  color={theme.colors.secondary}
+                />
+              </View>
+              <View style={styles.View_CardContentContainer}>
+                <MaterialCommunityIcons
+                  name="discord"
+                  size={scale(100)}
+                  color={theme.colors.primary}
+                />
+              </View>
+            </TouchableOpacity>
+          </CardContainer>
         </View>
-      </SafeAreaView>
-    );
-  } else {
-    // If chats haven't loaded, show indicator
-    return (
-      <View
-        style={[
-          { backgroundColor: theme.colors.background },
-          styles.loaderContainer,
-        ]}
-      >
-        <ActivityIndicator
-          size="large"
-          color={theme.colors.primary}
-          style={styles.loader}
-        />
       </View>
-    );
-  }
+    </SafeAreaView>
+  );
 };
 
 const styles = StyleSheet.create({
@@ -223,20 +89,10 @@ const styles = StyleSheet.create({
     backgroundColor: dozy_theme.colors.medium,
     flex: 1,
   },
-  View_HeaderContainer: {
-    backgroundColor: dozy_theme.colors.medium,
-    flexDirection: 'row',
-    padding: scale(14),
-    paddingTop: 0,
-  },
-  View_ChatNameContainer: {
+  headerContent: {
     flexDirection: 'column',
-    marginLeft: scale(10),
-    marginTop: scale(-2),
-    justifyContent: 'flex-start',
-  },
-  KbAvoidingView: {
-    flex: 1,
+    justifyContent: 'space-between',
+    alignItems: 'flex-start',
   },
   View_ContentContainer: {
     justifyContent: 'flex-start',
@@ -244,29 +100,27 @@ const styles = StyleSheet.create({
     paddingHorizontal: scale(10),
     paddingTop: scale(10),
   },
-  Img_Profile: {
-    width: scale(40),
-    height: scale(40),
-    borderRadius: 500,
+  View_CardHeaderContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
   },
-  Text_CoachName: {
-    color: dozy_theme.colors.secondary,
-    fontSize: scale(15),
+  View_CardContentContainer: {
+    height: 150,
+    justifyContent: 'center',
+    alignItems: 'center',
   },
-  Text_CoachTitle: {
+  Text_CardTitle: {
     color: dozy_theme.colors.secondary,
+  },
+  Text_CardSubtitle: {
+    color: dozy_theme.colors.secondary,
+    opacity: 0.5,
     marginTop: scale(-5),
+    marginBottom: scale(12),
   },
-  Text_FAQ: {
-    color: dozy_theme.colors.secondary,
-    marginBottom: -7,
-  },
-  faqButton: { flex: 1, alignItems: 'flex-end' },
-  loaderContainer: { flex: 1 },
-  loader: {
-    width: scale(45),
-    height: scale(45),
-    marginTop: '45%',
+  Icon_Clipboard: {
+    margin: scale(50),
     alignSelf: 'center',
   },
 });
